GradProject/requirements.txt
pandas==2.2.3
torch==2.5.1+cu121
torchvision==0.20.1+cu121
opencv-python==4.12.0.88
transformers==4.52.4
huggingface_hub==0.36.0
requests==2.32.5
Flask==3.0.0
python-dotenv==1.0.1
language-tool-python==2.8